src/query_map/query_map.rs
use std::sync::Arc;
use tokio::time::Duration;
use crate::{
    rpc::RpcClient,
    types::{Address, Balance},
    error::CommunexError,
};
use super::QueryMapConfig;

pub struct QueryMap {
    client: Arc<RpcClient>,
    config: QueryMapConfig,
}

impl QueryMap {
    pub fn new(client: RpcClient, config: QueryMapConfig) -> Result<Self, CommunexError> {
        Ok(Self {
            client: Arc::new(client),
            config,
        })
    }

    pub async fn get_balance(&self, address: &str) -> Result<Balance, CommunexError> {
        unimplemented!("get_balance not implemented")
    }

    pub async fn get_balances(&self, addresses: &[&str]) -> Result<Vec<Balance>, CommunexError> {
        unimplemented!("get_balances not implemented")
    }

    pub async fn get_stake_from(&self, address: &str) -> Result<Vec<Address>, CommunexError> {
        unimplemented!("get_stake_from not implemented")
    }

    pub async fn get_stake_to(&self, address: &str) -> Result<Vec<Address>, CommunexError> {
        unimplemented!("get_stake_to not implemented")
    }

    pub fn cache_stats(&self) -> CacheStats {
        CacheStats {
            refresh_count: 0,
        }
    }
}

pub struct CacheStats {
    pub refresh_count: u64,
} 